Supply histogram data to job metrics page
